main/tar: do not symlink tar under /usr
busybox's tar is installed only under /bin. This file being duplicated
across the /usr hierarchy breaks the /usr merge script. There's the
unlikely chance that something has tar hard-coded with the /usr path,
but there seems to be nothing in aports or mkinitfs, and that would
still be broken with tar from busybox