community/atools: split-off apkbuild-lint, favor atools-go
This project is not maintained anymore. For apkbuild-lint, there's
atools-go, but other linters / scripts and documentation is not
available anywhere else.

Split off apkbuild-lint into a subpackage and make it provide a virtual
`apkbuild-lint` package.

atools-go provides this virtual package as well, but with a higher
priority. This means apkbuild-lint from atools-go is preferred, but the
one from this package can still be explicitly installed if desired.