main/file: Import patch file instead of downloading it from GitHub
As noted by Celeste at [0], it's not recommended to use URL-based patches in
stable branches. Since the recently added patch is very small, we can simply
import it into the repository instead of downloading it.

Please note that the checksum has changed because the patch was downloaded with
the '?full_index=1' option.

[0]: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/77814#note_470256