community/flatpak: fix crash when using flatpak --user
When trying to use `flatpak --user`, it crashes with the following
error:

    error: While opening repository /var/lib/flatpak/repo: opening repo: opendir(/var/lib/flatpak/repo): No such file or directory

Two specific directories and a config file must be present, but nothing
creates them.

Include these directory along with a minimal configuration in this
package, so that --user works out of the box.

See: https://github.com/flatpak/flatpak/issues/5488