community/hexagonrpcd: improvements to how service is configured
This drops the /etc/conf.d/* config from the openrc subpkg. Upstream
states that the dir specified here is only for testing and that
distros/admins should install config for device-specific paths. By
having this default config in /etc/conf.d, it's more difficult for
downstream (pmOS) to do this, we have to use `replaces=` everywhere we
need to change this conf for the many different devices that need it.

This also adds support for loading the fw path from
/usr/share/hexagonrpc. pmOS can use this to install device-specific
config into (and this will make it nicer to re-use this config on
non-openrc init systems). The openrc init scripts for hexagonrpcd will
use the path in /etc/conf.d if a sys admin/user installed it, then it'll
try to load the path from /usr/share, and finally use the default value
for that fw path if it fails to find any in those config files.

Functionally there should be no impact over what path is used today.
This change is done primarily to support downstream packaging of config
for this service.