community/libtorrent: enable check again by enabling instrumentation
The failing tests rely on instrumentation and fail without it.
Instrumentation was disabled in 9f54084b7cdcb4fef4277eb50b124b35619a23dc
because it caused a build failure on x86_64. Presently, I can't
reproduce the build error on x86; hence, we might as well enable
instrumentation again (other distros leave it enabled too, e.g. Arch).

As an alternative solution, if instrumentation causes build issues in
again in the future, we could disable tests in LibTorrent_Test_Torrent_Utils
if LT_INSTRUMENTATION is not defined.