testing/ukify: add kernel-hook subpackage
The existing hooks in the secureboot package don't work with
systemd-efistub due to a mismatch in some hardcoded magic numbers.

Add a kernel hook for ukify, which works with systemd-efistub.

This new hook changes the default location and filename to comply with
systemd's conventions, so administrator will need to reconfigure their
systems' firmware to boot from these new paths.

A wiki entry with full details on configuring this is pending.

Fixes: https://gitlab.alpinelinux.org/alpine/aports/-/issues/16917