main/fish: ensure that tmux is not installed during check()
There are several racy tests which use tmux to send keystrokes to
the application holding the control terminal. These tests are
inherently racy and known to fail on our builders. Therefore, we
must ensure that tmux is not installed during check() to ensure
that these tests are skipped.

Otherwise, this causes hard-to-debug test failures.

See #16950