main/linux-lts: move aarch64 QRTR to lts
The commit b6ca5beb0162 (main/linux-lts: enable QRTR on x86, x86_64 and
aarch64) enabled QRTR on x86* lts kernel but aarch64 virt kernel which
appears to be a mistake.

It also enabled CONFIG_QRTR_MHI=m but did not enable the depndencies,
which means that it was effectively never enabled.

Clean this up so the lts aarch kernel has QRTR and only has the modules
that actually gets compiled so the config corresponds with reality.