main/gettext: upgrade to 0.24
gettext 0.22 has test failures when built against libunistrings 1.3 as
acknowledged by upstream. This has been fixed in 0.23.

libxml2 was upgraded to 2.14, but later reverted again because it
introduced a bunch of breakages against various packages, one of which
is gettext. Add a version constraint to make sure the builder picks op
2.13 instead of 2.14.

Fixes #17035