community/actionlint: fix linter test
According to `testdata/format/README.md`, actionlint.version value
should be an empty string when running unit tests. The template of
the expected output still inserts a version value, while in the parsed
template, the value is set to an empty string, the mismatch resulting in
`TestLinterFormatErrorMessageInSARIF` test failure at
`linter_test.go:463`.

The workaround replaces the version value in `sarif_template.txt` with an
empty string.