testing/ko: disable gobuilds tests
The tests fail since go 1.24. It effectively runs `go version -m <file>`
on a text file. In go 1.23 it exits with 0, while in go 1.24, it exits
with 1.

Running `go version` on a text file does not really make sense. The
command is meant to be used with go binaries. Disable the test for now
to let the builder continue.