testing/dumb_runtime_dir: make it work out-of-the-box
By changing the default parent directory in which XDG_RUNTIME_DIRs
are created from /run/user (which doesn't exist by default and needs
to be created by the user *with the right permissions*) to /run which
exists.

So now the default is `/run/$UID` we could also patch the code to
call the directory `/run/user-$UID`, if desired.