community/restic: disable node_xattr_all_test again
This was disabled prior to the upgrade to restic 0.18.0, however, with
this release the file was renamed. I incorrectly assumed that the file
was removed. Further, the test passes on the CI (presumably because /tmp
is not a tmpfs there) and hence I somehow missed this.

This is a fixup for 002e712b626537f3d7e9888f47ba0fa1d2653ada.