main/linux-lts: Enable Support for the ZYNQ MPSoC
Overall this change is +16K on the vmlinuz file and -12k on the modloop. It
enables all the kernel features required to support netbooting on Zynq MPSoC
platforms, which I have tested on an RFSoC4x2

CONFIG_SERIAL_XILINX_PS_UART:

Despite it's name CONFIG_SERIAL_XILINX_PS_UART enables support for a generic
Cadence UART. Building in this driver and enabling console support for it
makes bringup on zynqmp platforms and anything else that uses the Cadence
UART significantly easier.

My testing indicates that this increases the compressed size of the kernel
by 8K and decreases the size of the modloop by 12K which seems like a
reasonable trade to me given this is something you would want to have in
early boot

CONFIG_COMMON_CLK_ZYNQMP:

I've changed this to Y, on the Zynq MPSoC platform all the clocks are
managed by a Microblaze core. In order for linux to reconfigure the UARTs
and ungate the clock to the ethernet MAC this needs to be enabled.

Signed-off-by: Aled Cuda <aled@ucsb.edu>