testing/usbmuxd: split udev rules into subpackage
The provided udev rules automatically start the service when a
compatible device is detected, and attempt to stop it when all devices
are unplugged.

There are no reports of this working on Alpine; this automagically
started service does not seem to work properly. Specifically, it
won't allow interacting with iOS devices, which it the sole purpose of
this package.

Starting the service manually (or via OpenRC) does work. When
this rule is present, the two instances seem to race each other, and the
manually started usbmuxd only works _sometimes_. A manually started
instance works every time with the udev rule is not present.

Split the udev rule into a sub-package, allowing those who want to
experiment using it to opt-in, while getting it out of the way for the
typical use case.

---

Footnote: I actually run usbmuxd building the latest master from source,
since the latest stable release isn't compatible with devices from the
last several years. Other tools pull usbmuxd as a dependency, and the
rule interferes even in that case.