community/py3-wgnlpy: fix exception when WGPEER_A_LAST_HANDSHAKE_TIME is None
The value of a peer's last handshake NLA can be None which triggers an TypeError
exception inside wgnlpy. This breaks ifstate from time to time when it is used
to configure wireguard VPNs and this bug is triggered. Although a PR[1] was
opened longer time ago the issue had not been fixed upstream, yet.

[1] https://github.com/ArgosyLabs/wgnlpy/pull/5