testing/meli: upgrade to 0.8.11
Upgrade to 0.8.11.

While at it, also reduce the size of the patch and split it up into
separate patches. This hopefully makes maintaining the patches are
easier. The patches make it so that libz isn't statically linked, the
"cli-docs" feature for manual pages with the CLI is disabled, and the
exposed subcommands for the "cli-docs" feature are hidden because
they're not available anyway.

The code for D-Bus notifications and for displaying the manual pages
isn't removed like before because this code is conditionally compiled in
only when the features "dbus-notifications" and "cli-docs" are enabled.
This shrinks the size of the patch significantly and shouldn't affect
the size of the binary.

Signed-off-by: Thomas Böhler <witcher@wiredspace.de>