community/pmbootstrap: depends on cmd:kpartx
Depend on the split package kpartx, so that pmbootstrap can be installed
in scenarios where multipath-tools or one of its dependencies is in
conflicts with other packages (or simply undesirable).

Allows installing pmbootstrap on hosts which are running mdevd instead
of eudev.

Tested by building an image and also running a qemu image.