testing/gummiboot: move from main
Gummiboot is dead, it was absorbed by systemd a long time ago, and
several users reported that doesn't work since Alpine v3.21 (see #16691).
I guess it's time to say goodbye. :( That said, it works on my machine
after rebuilding, so let's keep it in testing for now, in case anyone
needs it.

We have two alternatives: systemd-efistub and stubbyboot-efistub.