community/cyclone: run tests with just-built icyc
instead of icyc (potentially) from cyclone-stage0

There are 2 providers of cmd:icyc, cyclone-stage0 and cyclone.
Tests were run with icyc from $PATH, which could be from either
of these 2 aports. Using ./icyc forces tests to be run with
the icyc executable that has just been built, which is the
one that will end up in cyclone (as we are building cyclone).

This avoid running the icyc from cyclone-stage0, because it seems
https://gitlab.alpinelinux.org/alpine/aports/-/issues/16657 is only reproducible when bootstrapping from -stage0.
However, edge self bootstraps from cyclone, and it is only while
preparing stable releases that -stage0 comes into the picture,
therefore we must remember to explicitly switch to -stage0 when
looking into issues that are encountered on the stable builders.

To further complicate matters, when cyclone-stage0 is built also
matters. According to https://pkgs.alpinelinux.org/packages?name=cyclone-stage0&branch=edge&arch=
cyclone-stage0=0.36.0-r0 was built on 2024-03-01 for 7 archs with
the exception of loongarch64 that was built on 2024-09-08, switch
cyclone in edge to bootstrap from -stage0, and it will pass for
the 7 archs that built -stage0 on 2024-03-01.