community/cyclone-stage0: apply gcc 14 feature detection patch from cyclone
To answer my own question raised 2 commits prior, the reason icyc
from cyclone-stage0 fails to run cyclone tests while the one from
cyclone works is because this patch was only applied to cyclone.

So, applying it here should solve #16657 for good.

Also, with cyclone now fixed, the 3.22 builders
for aarch64, armhf and x86 should be unblocked.