community/proftpd: add mod_vroot to the list of mods
mod_vroot is an essential module for proftpd, which will hide the parent
path from the user connecting to the FTP. This is usually what most of
the users want. Because of that, this module is added in most of all
other distributions in the base package.

This commit will add mod_vroot as a subpackage.