testing/biboumi: move from community
Since Botan2 reached End of Life it will not be part of our 3.22
release. Instead of removing it entirely we move it to testing/ until
dependants can catch up and switch to Botan3 or other, therefore we move
this to testing/ for now as well.

The upstream project has also moved, so we update url and source to
reflect that.