main/gzip: Don't create symlinks back to /usr/bin
Creating these symlinks from /bin to /usr/bin results in broken symlinks
on usr-merged installs. Busybox symlinks are in /bin, so installing gzip
and gunzip to /bin is enough to overwrite them.

AFAIK busybox and gzip (this pkg) are the only things providing these
cmds, so overwriting the busybox symlinks should be enough to ensure
that GNU gzip/gunzip is used.