main/curl: Use more common ca-bundle path
!69012 explicitly added the `--with-ca-bundle` argument and used
`/etc/ssl/cert.pem` as the path. This is a symlink to the actual file
`/etc/ssl/certs/ca-certificates.crt`. This is also the path which is
auto-detected by `curl` when the argument is not provided.

It would allow static binaries built on Alpine to be used on other
distributions correctly, since most distros seem to have a certificates
bundle or a symlink at the `/etc/ssl/certs/ca-certificates.crt` path.