testing/linux-stable: package vmlinux.h for BPF development
x86_64 and aarch64 versions of this kernel are enabled now, other archs
might be missing config for this to work correctly and can be enabled
later.

I investigated using the vmlinux.h target in tools/bpf/bpftool/Makefile
to generate this but couldn't get it to work with the way we are
building the kernel (in a subdir under the tree root), the makefile has
a relative import. Calling bpftool directly does the same thing.