main/mpfr4: adjust logic for forming patch sources
Before this change, when the pkgver is not a patched version, the while
loop for adding the patches is run with an invalid loop control. If the
pkgver is unpatched version (as it is now), the control statement for
the while looks like this:

	while [ $_i -le 4.2.2 ]; do
		...
	done

In this case, '4.2.2' (the package version) is invalid for the '-le'
test operator. Adding the outer if statement prevents this.

I didn't bump the pkgrel since there is no actual functional change to
the package output; this only prevents a shell error from occuring.