community/gthumb: remove gstreamer subpackage
Currently, gthumb-gstreamer cannot be installed, since it both depends
on the main gthumb package, as well conflicts with it:

    $ doas apk add gthumb-gstreamer
    ERROR: unable to select packages:
      gthumb-3.12.8.1-r0:
        conflicts: gthumb-gstreamer-3.12.8.1-r0[so:gthumb:libslideshow.so=0]
    	       gthumb-gstreamer-3.12.8.1-r0[cmd:gthumb=3.12.8.1-r0]
        satisfies: world[gthumb] gthumb-gstreamer-3.12.8.1-r0[gthumb=3.12.8.1-r0]
      gthumb-gstreamer-3.12.8.1-r0:
        conflicts: gthumb-3.12.8.1-r0[so:gthumb:libslideshow.so=0] gthumb-3.12.8.1-r0[cmd:gthumb=3.12.8.1-r0]
        satisfies: world[gthumb-gstreamer]

libslideshow.so can be easily resolved by adding it to somask, but
conflict on cmd:gthumb is more complicated. This wasn't a problem
before, it seems that apk's solver has changed.

As discussed in !85657, when gthumb is combined with any desktop
enviroment in which it is useful, it already pulls in most (or all) of
the dependencies gstreamer would. Thus trying to split it into a
subpackage is not worth the trouble.

Resolves #17617

See !85657 !91662

Co-Authored-By: Jakub Jirutka <jakub@jirutka.cz>