community/linux-stable: x86_64: Fix Bay Trail and Cherry Trail PMIC configuration
The Intel Bay Trail and Cherry Trail SoCs used in many x86 tablets come
with a more ARM / embedded-hw like PMIC setup. The PMIC's resources are
accessed early on when the ACPI tables are being parsed the first time.
Both through ACPI opregion calls as well as through PMIC GPIO accesses.

To ensure that these PMICs are available early enough all necessary
drivers for accessing the PMICs and their opregion and GPIO drivers
need to be builtin.

For everything to properly work on these tablets the following changes
are necessary:

1. Change I2C_DESIGNWARE_PLATFORM from m to y and enable
   I2C_DESIGNWARE_BAYTRAIL which offers arbritration between the kernel
   accessing the PMIC and the SoC's punit accessing it (both share the
   same I2C controller).

2. The devices which I2C_DESIGNWARE_PLATFORM binds to are child devices
   of the LPSS island, change MFD_INTEL_LPSS_* from m to y to make sure
   these are immediately available at boot.

3. Now that I2C_DESIGNWARE_PLATFORM=y, various CONFIG_INTEL_SOC_PMIC_*
   options, which depend on this, become available. Enable these boolean
   options. This enables MFD drivers which create child devices for
   various functions of the PMIC. This in turn makes a bunch of Kconfig
   options for drivers for these child devices avaialable.

4. Set the now available *_PMIC_OPREGION boolean options to y

5. Enable the drivers for all PMIC child devices, using y for the GPIO
   drivers and m for the rest.

Note this brings the config inline with Fedora which has seen a lot of
testing on these kind of devices.

This fixes a whole bunch of things, e.g. backlight control on many
Bay Trail devices and fuel-gauge (battery monitoring) on a a bunch
of Cherry trail devices and many many more things.