community/libudev-zero-dev: provide eudev-dev
libudev-zero works as a drop-in repleacement for eudev. In the same way,
libudev-zero-dev works as a drop-in replacement for eudev-dev.

eudev-dev depends on eudev-libs=$pkgver-$pkgrel. The version constraint
prohibits eudev-dev from being installed on systems with libudev-zero
installed, despite the latter including provides=eudev-libs.

Add provides=eudev-dev to libudev-zero-dev, so that it can be used as a
drop-in replacement too. This allows systems with libudev-zero installed
to properly resolve build dependency for ports which rely on eudev-dev