community/linux-stable: x86_64: Make simpledrm driver builtin
On EFI systems without a fb or drm driver being builtin there is no way for
the kernel to show messages until the GPU driver (e.g. i915.ko) loads this
means that if there are some issues with the initrd before the GPU driver
loads nothing is shown to the user.

Make the simpledrm driver builtin so that early kernel messages / errors
will be visible.

Note the x86_64 config already has CONFIG_FB_EFI=y but that only comes into
play if drivers/firmware/sysfb.c fails to register a new simplefb/simpledrm
compatible framebuffer platform-device in which case it will fallback to
registering a "efi-framebuffer" platform-device to wich the efifb driver
binds. In practice this almost never happens so simpledrm is needed to show
early boot messages.

Signed-off-by: Hans de Goede <hansg@kernel.org>