community/xterm: build with --enable-block-select
XTerm 397 added support for selecting a rectangular block of text
(https://invisible-island.net/xterm/xterm.log.html#xterm_397).

I have, since then, tested the patch on all versions up through 401, and
found no problems. Though it is a little crude, it is a welcome feature.

This is already included, all we need is to enable the build flag as I
have done here with `--enable-block-select`.