main/linux-lts: x86_64: Backlight config updates
Enable the LP855X and Dell UART backlight drivers which are used on some
x86_64 devices. Disable the BACKLIGHT_PWM and BACKLIGHT_ADP8860 options
which are not used on any x86_64 devices

Remove all CONFIG_LCD* options these are only for use with obsolete fbdev
drivers which are used on old non x86 embedded platforms and which are not
enabled in the x86_64 config.

Remove the BACKLIGHT_PCF50633 option, this does not exist in current
lts kernels.

Signed-off-by: Hans de Goede <hansg@kernel.org>