main/alsa-utils: Make regex to extract audio card numbers more robust
The current regex matches the digit **zero** or more times. Therefore,
it also matches lines which do not contain a digit. This causes the
service to wait (with sleep 2) for cards that don't even exist. Further,
it prevents the restore command from failing for these non-existent
cards.

This patches fixes this issue by requiring **one** or more digits in
the `/proc/asound/cards` lines. While at it, I also fixed the local
declaration for the loop variable.