community/forgejo: revert zstd compression level on armhf/v7
There appears to be a bug in zstd on armhf/v7 that causes the build to panic
due to memory exhaustion. This issue does not appear to happen in docker, so CI
does not catch it.

Upstream [ran into similar issues][0], but they seem to have solved it, since
they changed back to zstd.SpeedBestCompression.

Revert it on those arches for now to fix the builders. This would result
the filesize to increase about ~2MB according to upstream.

[0]: https://codeberg.org/forgejo/forgejo/issues/8165