community/mimalloc2: build with -DMI_NO_OPT_ARCH=ON
As per #17395 running things built against `mimalloc2`
on a Raspberry Pi (or indeed any armv8.0 aarch64 arch)
causes an `Illegal instruction`. This was noticed for me
with `vaultwarden`.

This commit adds the `-DMI_NO_OPT_ARCH` cmake option, which
has the effect of not setting the `-march=armv8.1-a` on
armv8 architectures.  The RPI 4 in particular uses a
Cortex A-72 core, which is `armv8-a`, not 8.1.

Ths option ha no effect on other architectures and so
does not need to be set conditionally.

References:

* Disussion on the mailing list: https://lists.alpinelinux.org/~alpine/users/%3C20250725094929.2abbb421%40baryte%3E
* Issue on mimalloc2 tracker: https://github.com/microsoft/mimalloc/issues/1095
* Link to relevant lines of upstream CMakeLists.txt: https://github.com/microsoft/mimalloc/blob/d2e259cb57266e316b242aed7441641908c678c3/CMakeLists.txt#L442
* Link to architecture of Cortex cores (pdf): https://documentation-service.arm.com/static/67bf3e13ce2747241fcea4ca?token=

Fixes #17395 (https://gitlab.alpinelinux.org/alpine/aports/-/issues/17395)