main/systemd-boot: increase number of stub sections for ARM
The default is 30, and we (pmOS) may want to add many more than 30
dtbauto sections when building a UKI in the future for more 'generic'
UKIs that can multiple devices.

3000 is kinda arbitrary, the linux-next kernel we have in pmOS ships
1.6K dtbs, so 3K seemed like it would give us more room to grow. On the
other hand, this increases the stub size a little:

30 extra sections:
81K Jul 27 05:04 linuxaa64.efi.stub

3000 extra sections:
197K Jul 28 20:39 linuxaa64.efi.stub

This seems trivial to me. But if someone has a problem with it we can
reduce the number to 2K.