community/kicad-library: patch footprints to use wrl
The 3D models in kicad-library-3d contain both wrl and step files for
each 3D model. However, when KiCad is not linked against opencascade,
only the wrl models can be viewed. So just patch the footprints to
always refer to the wrl file that can be viewed both with and without
opencascade.