community/java-common: fix missing symlink when the jvm is empty
In case a jvm is uninstalled, its folder keep some empty folders. This
can trick the trigger script to link it as "default-jvm", and produce
no /usr/bin/java symlink at the end.

Checking that the bin/java file is there seems enough to be sure the jvm
folder is sane enough.