community/mate-utils: split dictionary into subpackage
mate-dictionary is just 6MB, but installing mate-utils plus its
dependencies is a total of 144MB.

Split mate-dictionary into a subpackage, allowing the installation of
mate-dictionary without the rest of mate-utils and its dependency tree.

Make mate-utils depends on mate-dictionary, so that installing
mate-utils still installs mate-dictionary. This ensures that existing
installations are not affected by this change.