community/linux-stable: aarch64: enable necessary configs for x1e80100
this makes linux-stable bootable on a Lenovo Yoga Slim 7x Gen9
Snapdragon laptop, although it also requires bootloader configuration to
load the correct dtb as well as adding a bunch of kernel modules to the
initramfs.

while this list of modules/configs makes the device bootable, it is not
exhaustive; other options may need to be enabled later to support more
hardware features.