testing/idevicerestore: ugrade to 1.0.0_git20250914
The latest tagged release is 5 years old and doesn't work with the
latest Apple devices, but the latest in git does.

This adds a dependency on usbmuxd, which is required for mode switching
when connected to macbooks (and maybe other things?)

This also adds a udev rule so that users in the plugdev group can run
this tool on devices in DFU mode without requiring root privileges.