scripts/mkimg.*.sh: readd riscv64 to standard
It is already possible to install Alpine with EFI bootloader on riscv64,
but it needs to be done with the generic U-Boot tarball, booting with
which is not trivial.

There's a LTS kernel for the standard profile now, so it's possible to
return riscv64 to the list of platforms with ISO support. Just add the
relevant clauses to profile_standard and profile_standard; the list of
initfs_features was checked against the existing U-Boot profile.

This reverts commit 877e7b1b61aa6f873defb9c453b788cb9d126a21.

Fixes #17531