community/zoneminder: downgrade to 1.36.36
because 1.37.63 is vulnerable to CVE-2024-51482 but version tags are not
maintained consistently for the 1.37 series. This reverts commits
65c8fabe7d67c4e1efe20ba891de408358e0c60b and
574d9f47a15ce99c9d25e0c9804e6ebd2b5c26d3. See also commit
2d04904554f159d3c8e3d5eef186331c8cc377f6.

The dependencies are adjusted such that the C++ code uses FFmpeg 7 while
the Perl and PHP code invokes the FFmpeg 8 executable. We can fully
migrate to version 8 when a stable 1.38 release comes out.