main/alpine-baselayout: do not create usr/lib/firmware
This was added during the /usr-merge basically copying the directories
around. However, contrary to the other usr/lib paths, no firmware can
ever be installed under usr/lib until #17423 is resolved (unless the
symlink is already in place, in which case creating the directory is
unnecessary). Additionally, compared to the other paths, the
lib/firmware path may be a symlink in some situations (see #17580),
breaking the /usr-merge script. Therefore, remove this.