community/syncplay: add -client-gui, -server-openrc and -server-nftrules
The openrc package install the initd and confd service files, and
setup a default salt and password.

The nftrules packages is installed automatically with nftables-rulesets.
It provide a simple nftrules hole for the default port in a file located
in /usr/share/nftables.avail/.

The -client-gui wrap the dependencies to display the window. The client
side can run without any extra gui using --no-gui.