community/mycli: patch out llm requirement
I'm not interested in packaging and maintaining py3-llm and it seems
no one else is either, yet.

Either way, I don't think we should have a hard dependency on such a
package.  Preferably it would be optional, the functionality wouldn't
have to be patched out like this but to use it you would have to add the
py3-llm package to your system.