community/litecli: patch out llm requirement
We have not yet packaged a py3-llm aport plus additional required
dependencies.

I don't think we, trying to keep the distribution small, should have a
hard dependency on such a package.  Preferably it would be optional, the
functionality wouldn't have to be patched out like this but to use it
you would have to add the py3-llm package to your system.