main/chrony: move vrf handling to init script
the RC_VRF_EXEC support was not upstreamed to openrc, so do not rely on
it.

this implementation is slightly different from the openrc patch, in that
the daemon supervisor does not run under the designated vrf, whereas
with RC_VRF_EXEC it did.

References: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/91552