community/nmon: use full path in prepare()
The working directory when in the `prepare()` function defaulted to
`$startdir` when `$builddir` did not yet exist. Use full path in the
`cp` command within `prepare()` to ensure the source file is found
independent of the current working directory.

```
/var/cache/distfiles/v3.23/lmon16q.c: OK
cp: can't stat 'lmon16q.c': No such file or directory
>>> ERROR: nmon: prepare failed
```