community/bird_exporter: drop explicit bird dependency
The package has an explicit dependency on the bird package that
prevents installling it with bird2. As it does not require any
bird binaries but just the bird socket (i.e. when bird is run
inside a container) there is no need to depend on it at all. This
makes the package work with both bird variants.