community/hugo: fix checksum mismatch
The checksum of the upstream tarball had not changed, but use of a
corrupted cached tarball resulted in a checksum mismatch. Temporarily
rename the package filename to trigger re-fetching of the tarball.

```
sha512sum: WARNING: 1 of 1 computed checksums did NOT match
Because the remote file above failed the sha512sum check it will be renamed.
Rebuilding will cause it to re-download which in some cases may fix the problem.
Renaming: hugo-0.152.2.tar.gz to hugo-0.152.2.tar.gz.7fed73ed
```