community/haruna: fix checksum mismatch
The checksum of the upstream tarball had not changed, but use of a
corrupted cached tarball resulted in a checksum mismatch. Temporarily
rename the package filename to trigger re-fetching of the tarball.

```
/var/cache/distfiles/edge/haruna-1.6.0.tar.xz: FAILED
sha512sum: WARNING: 1 of 1 computed checksums did NOT match
Because the remote file above failed the sha512sum check it will be renamed.
Rebuilding will cause it to re-download which in some cases may fix the problem.
Renaming: haruna-1.6.0.tar.xz to haruna-1.6.0.tar.xz.bc65a64f
```