community/lightdm: remove debian specific check from lightdm.service
During testing for v25.12 of postmarketOS, it was discovered that
`lightdm.service` contains a debian specific check if `lightdm` is set
as default displaymanager in `/etc/X11/default-display-manager`.

As Alpine (and pmOS) do not use this file, the check for it fails and
thus lightdm does not gets properly started.

cc @fossdd

Signed-off-by: Robert Eckelmann <longnoserob@gmail.com>