main/apk-tools: disable zstd again
zstd increases the base image with ~700k and is apparently not used by
anything at all.

zstd was first introduced in commit 4730112da306 (main/apk-tools:
upgrade to 3.0.0_rc5_git20250613) without consideration of the
dependency graph.

The support was lost before alpine 3.23.0 release in commit c61847939790
(main/apk-tools: upgrade to 3.0.0) so Alpine Linux 3.23.0 was released
without zstd and apparently nothing broke due to the lack of it.

It was again re-enabled in commit 5fd5ae2c12c1 (main/apk-tools: build
against zstd), again without any explanation why it is needed for.

Remove it until we have a clear reason why we need it in our base image.

This reverts commit 5fd5ae2c12c1 (main/apk-tools: build against zstd)