community/linux-stable: Enable CONFIG_CORE_DUMP_DEFAULT_ELF_HEADERS
I have run into problems matching a core dump to the right binary. The
issue is that without CONFIG_CORE_DUMP_DEFAULT_ELF_HEADERS, there is
nothing to identify the exact binary that caused the dump.

With CONFIG_CORE_DUMP_DEFAULT_ELF_HEADERS, the kernel includes the
first page of the executable and each shared library, and gdb checks
that it matches a given binary.

This can be changed at runtime, in /proc/<pid>/coredump_filter. The
config is just a default.