main/lua5.5: make readline a soft dependency
Lua 5.5 introduces support to open readline with dlopen, so you no
longer need readline to execute your srcipts, but can still get a nice
interactive Lua when readline is installed. Enable this feature so
readline only is a soft dependency.

Upstream uses libreadline.so, which is unacceptable for us. Instead,
extract the soname from the library and use that.