community/obex-data-server: move from main
gdk-bixbuf will be moved to community too, and obex-data-server uses it
for "BIP support". The alternative provider for BIP support is
imagemagick which is also in community. Or BIP support can be disabled
if obex-data-server should be kept in main, but the project is
unmaintained and no other dependencies on o-d-s are in main.

Part of https://gitlab.alpinelinux.org/alpine/aports/-/issues/17550