main/busybox: explicitly return 0 in loadkmap stop action
If we use return without an argument it will use the exit code of
the last command. For some reason, this is non-zero with current
OpenRC which causes the stop() action to fail.

However, since we want to overwrite OpenRC's default_stop() to signify
that this service doesn't have any meaningful stop() action, we don't
want to delete the function entirely. Instead, just always return a
zero exit status.

While at it, do the same in start().

Fixes #17850